Phase 5 commit 3: extract notes/hooks/{useSaveSync,useFolderTree}.ts
- useSaveSync owns the savingState/savedAt slots and the 2 s debounced
  auto-save effect (38 lines of effect + complex deps). Returns the
  state and setters so other Notes.tsx code that flips files dirty or
  reads saving status keeps working unchanged.
- useFolderTree wraps the 7-line useMemo over folders × folderFiles.

Notes.tsx 1108 → 1064 lines. Build clean.
